requirements.txt
dj-database-url==0.5.0
Django==1.11.23
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
mysqlclient==1.3.4
psycopg2==2.8.3
pytz==2019.2
whitenoise==4.1.3
